chore(hook): added useTheme hook
